requirements.txt
imageio==1.5
numpy==1.11.1
Pillow==3.3.1
apistar==0.5.33
gunicorn==19.8.1
